sort client list when it is created
this isn't ideal, but it's gonna work for now. eventually, I probably want the turns to be in order of who joined the lobby first, not who connected first.

additionally, this will probably need to be reworked later when leaving/joining midgame is implemented, but I think that's a long time away.